refactor: mypage styled component 정리
